# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG World - Hệ thống nhập vai chiến đấu NPC
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Character {
  id: number;
  name: string;
  class: string;
  level: number;
  hp: number;
  maxHp: number;
  xp: number;
  xpToNext: number;
  gold: number;
  /** @nullable */
  worldId: number | null;
  /** @nullable */
  equippedWeapon?: string | null;
  /** @nullable */
  equippedArmor?: string | null;
  createdAt?: string;
}

export interface CharacterInput {
  /** @minLength 1 */
  name: string;
  class: string;
}

export interface CharacterStats {
  characterId: number;
  attack: number;
  defense: number;
  speed: number;
  critRate: number;
  totalBattles: number;
  wins: number;
  losses: number;
}

export interface World {
  id: number;
  name: string;
  description: string;
  theme: string;
  minLevel: number;
  npcCount: number;
  isBossWorld?: boolean;
}

export interface Npc {
  id: number;
  name: string;
  type: string;
  level: number;
  hp: number;
  maxHp?: number;
  worldId: number;
  difficulty: string;
  xpReward?: number;
  goldReward?: number;
  isBoss?: boolean;
}

export interface Battle {
  id: number;
  characterId: number;
  npcId: number;
  status: string;
  currentTurn: number;
  characterHp?: number;
  npcHp?: number;
  /** @nullable */
  xpGained?: number | null;
  /** @nullable */
  goldGained?: number | null;
  log: string[];
  createdAt?: string;
}

export interface BattleStart {
  characterId: number;
  npcId: number;
}

export type BattleActionInputAction = typeof BattleActionInputAction[keyof typeof BattleActionInputAction];


export const BattleActionInputAction = {
  attack: 'attack',
  skill: 'skill',
  defend: 'defend',
  flee: 'flee',
} as const;

export interface BattleActionInput {
  action: BattleActionInputAction;
}

export interface BattleActionResult {
  battle: Battle;
  log: string[];
  isOver: boolean;
  /** @nullable */
  result?: string | null;
}

export interface InventoryItem {
  id: number;
  characterId: number;
  name: string;
  type: string;
  rarity: string;
  attackBonus?: number;
  defenseBonus?: number;
  isEquipped?: boolean;
}

export interface EquipInput {
  itemId: number;
}

export interface LeaderboardEntry {
  rank: number;
  characterId: number;
  name: string;
  class: string;
  level: number;
  totalBattles: number;
  wins: number;
}

export interface ProgressionSummary {
  totalCharacters: number;
  totalBattles: number;
  totalNpcsDefeated: number;
  topClass: string;
  worldsExplored: number;
}