# lib/api-zod/src/generated/types/battleActionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG World - Hệ thống nhập vai chiến đấu NPC
 * OpenAPI spec version: 0.1.0
 */
import type { BattleActionInputAction } from './battleActionInputAction';

export interface BattleActionInput {
  action: BattleActionInputAction;
}