# lib/api-zod/src/generated/types/battleActionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG World - Hệ thống nhập vai chiến đấu NPC
 * OpenAPI spec version: 0.1.0
 */
import type { Battle } from './battle';

export interface BattleActionResult {
  battle: Battle;
  log: string[];
  isOver: boolean;
  /** @nullable */
  result?: string | null;
}