# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG World - Hệ thống nhập vai chiến đấu NPC
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: number;
  name: string;
  class: string;
  level: number;
  hp: number;
  maxHp: number;
  xp: number;
  xpToNext: number;
  gold: number;
  /** @nullable */
  worldId: number | null;
  /** @nullable */
  equippedWeapon?: string | null;
  /** @nullable */
  equippedArmor?: string | null;
  createdAt?: string;
}