# lib/api-zod/src/generated/types/equipInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG World - Hệ thống nhập vai chiến đấu NPC
 * OpenAPI spec version: 0.1.0
 */

export interface EquipInput {
  itemId: number;
}